+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version log for Spekkio.
 - Added queries for Scrumify
 - Added POST handler wrapper
 - Added links to related repositories to README
+- Added latest version badge to README
 ### Modified
 - Enhanced Error response struct in models